@@ -2,16 +2,26 @@
   "name": "javascript-action",
   "version": "1.0.0",
   "description": "JavaScript Action Template",
-  "main": "index.js",
+  "main": "src/index.js",
+  "directories": {
+    "test": "test"
+  },
+  "type": "module",
   "scripts": {
-    "lint": "eslint .",
-    "prepare": "ncc build index.js -o dist --source-map --license licenses.txt",
+    "build": "npx tsc",
+    "lint": "eslint src/*.ts",
+    "prepare": "ncc build src/index.ts -o dist --source-map --license licenses.txt",
     "test": "jest",
     "all": "npm run lint && npm run prepare && npm run test"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/actions/javascript-action.git"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.(ts|tsx)$": "babel-jest"
+    }
   },
   "keywords": [
     "GitHub",
@@ -28,12 +38,19 @@
     "@actions/core": "^1.10.0",
     "@actions/github": "^5.1.1",
     "@octokit/core": "^4.2.0",
-    "jshint": "^2.13.6",
-    "octokit": "^2.0.14"
+    "node-fetch": "^2.6.7"
   },
   "devDependencies": {
+    "@babel/core": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@babel/preset-typescript": "^7.21.4",
+    "@jest/globals": "^29.5.0",
+    "@types/jest": "^29.5.0",
+    "@types/node-fetch": "^2.6.2",
     "@vercel/ncc": "^0.36.1",
-    "eslint": "^8.0.0",
-    "jest": "^29.5.0"
+    "eslint": "^8.38.0",
+    "jest": "^29.5.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4"
   }
 }
