@@ -1,10 +1,8 @@
-let wait = function (milliseconds) {
+export function wait(milliseconds: number|any): Promise<string> {
   return new Promise((resolve) => {
     if (typeof milliseconds !== 'number') {
       throw new Error('milliseconds not a number');
     }
-    setTimeout(() => resolve("done!"), milliseconds)
+    setTimeout(() => resolve("done!"), milliseconds);
   });
-};
-
-module.exports = wait;
+}
